Flyttet kode for å sjekke om api-queries og query-params er klare
